flashcards/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
google-generativeai>=0.9.0,<1.0.0
pypdf==5.0.1
python-docx==1.1.2
Pillow==10.4.0
pytesseract==0.3.13
